requirements.txt
nicegui
anthropic
mcp
python-dotenv
requests